evm_transfer_monitor/requirements.txt
sqlalchemy>=1.4.0
requests>=2.25.0
asyncpg>=0.27.0
aiohttp>=3.8.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
